Fixed bug in indicator system
Removed dancer object from playpoint script (VideoActivator). Set the initial layer of each dancer clip to 0 (default) upon enabling, so particle won't collide with it. Layer will be set to 3 (theater) inside IndicatorSystemController once the object becomes the target object so that particles can collide with it. Once a new target is set, the old target's layer will be set back to default.